fix: rename some files to work around a CMake bug in MSYS, refs #1776
The bug is triggered in MSYS when CMake tries to compile libraries and
LTO is enabled; in this case CMake 3.20 and earlier uses "ar cr" to
create and extend the libraries, which overwrites old entries in the
archive with the same basename. CMake should use "ar qr", but this is
fixed only in CMake 3.21.

See these issues for more details:

https://github.com/igraph/igraph/issues/1776
https://gitlab.kitware.com/cmake/cmake/-/issues/21988