remove local_scan_1_sumweights
igraph_i_local_scan_sumweights was faster, but incorrect. It didn't
count loops on neighbors, it counted loops on the starting vertex twice,
and it counted only one of the multiple edges between two neighbors. When
counting edge weights this is almost always incorrect because you need
some way of combining weights, and not just pick a random one. I didn't
really find a way to fix this.